update to use 256

+++ contracts/vaults/BaseVaults/RibbonThetaVault.sol
@@ -271,7 +271,7 @@ contract RibbonThetaVault is RibbonVault, RibbonThetaVaultStorage {
      * @notice Sets the new leverage factor
      * @param newLevFactor is the leverage factor to use
      */
-    function setLevFactor(uint16 newLevFactor) external onlyOwner {
+    function setLevFactor(uint256 newLevFactor) external onlyOwner {
         require(newLevFactor >= 10 ** 6 && newLevFactor <= 2 * 10 ** 6, "Leverage not 1x - 2x!");
         levFactor = newLevFactor;
     }
